Added: clicking on the bank's URL, bank's phone number, or country coordinates sends the user to an application that can process this data (browser, dialer, maps). Fixed issue with non Int coordinates.

+++ app/src/main/java/com/ordresot/binviewer/data/remote/dto/CountryDto.kt
@@ -6,6 +6,6 @@ data class CountryDto(
     val name: String?,
     val emoji: String?,
     val currency: String?,
-    val latitude: Int?,
-    val longitude: Int?
+    val latitude: Double?,
+    val longitude: Double?
 )
